Refactor: Move fetchCryptocurrencies dispatcher from Dashboard to App
It makes sense to put the dispatcher for getting cryptocurrency data in the root App component instead of the Dashboard because the Dashboard is not the only component in its level of the hierachy which uses the data. Futhermore, the App component only mounts once, making it trivial to call setInterval in a component lifecycle method.