Currency change feature
Checkbox for changing currency.
Actions and reducers for currency change and keeping track of the active refresh interval.
Currency change reducers set loading to true and cause a saga to fetch new results in the selected currency.
Keep track of the active refresh interval to prevent duplicate interval functions when the currency is changed.